refactor: move Writer into the library as dogrun::writer
Pure move of highlight(), ReadmeUpdater, Writer, and update_readme_fzf
from main.rs into src/writer.rs with pub visibility on the Writer API;
main.rs keeps only argument parsing and file IO. Writer was private to
the binary, which is why integration tests could not reach it and the
wezterm test had to re-implement what it meant to verify. Generated
output is unchanged (golden test).
