+++ generator/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod conv;
 pub mod highlight;
+pub mod writer;
